Move data to fixtures file
